Rename rewritedSearchData.json to searchData.json, fix the "rewrited" typo
- javascript/searchRewrite.ts -> javascript/search.ts
- javascript/searchRewriteTest.js -> javascript/searchTest.js
- writeRewritedSearchData() -> writeSearchData(), and its output file
  rewritedSearchData.json -> searchData.json

This collided with the older, unrelated searchData.json written by
generateSearchData.ts (a legacy search-index pipeline predating the
current one; a comment at its call site already noted "this is meant
to be temp; also, will remove the original generateSearchData after
the updation at the frontend is completed" - that migration is long
done). Removed that legacy pipeline entirely rather than picking a
different name, since nothing else imports from it (verified) and its
output was already unused/dead.

Verified by rebuilding json_js/json_py/json_scm from scratch: each
edition now writes a single searchData.json with the current
(indexTrie/textTrie/idToContentMap) structure, and every other
generated file is byte-identical to before removing the legacy
pipeline.

README updated to match (json/searchData.json, javascript/search.ts).

A companion change is needed in the frontend repo, which currently
fetches rewritedSearchData.json by that name for both SICP JS and
SICPy - filing that as a separate issue there.
