+++ README.md
@@ -9,7 +9,7 @@ This repository contains processing scripts and sources for *Structure and Inter
 * [PDF edition](https://sicp.sourceacademy.org/sicpjs.pdf)
 * [Interactive SICP JS](https://sourceacademy.org/sicpjs), backed by structured per-section JSON textbook content, e.g. [json/1.1.1.json](https://sicp.sourceacademy.org/json/1.1.1.json)
 * [Comparison edition](https://sicp.sourceacademy.org), Scheme vs. JavaScript
-* Search support: the [frontend](https://github.com/source-academy/frontend)'s Text Search and Index Search buttons fetch [json/rewritedSearchData.json](https://sicp.sourceacademy.org/json/rewritedSearchData.json) from this public static host (not the Source Academy backend) and filter it locally; generated by `yarn json` (`javascript/searchRewrite.ts`), which runs as part of `deploy-pages.yml` on every push to `master`
+* Search support: the [frontend](https://github.com/source-academy/frontend)'s Text Search and Index Search buttons fetch [json/searchData.json](https://sicp.sourceacademy.org/json/searchData.json) from this public static host (not the Source Academy backend) and filter it locally; generated by `yarn json` (`javascript/search.ts`), which runs as part of `deploy-pages.yml` on every push to `master`
 * Source code archive ([sicpjs.zip](https://sicp.sourceacademy.org/sicpjs.zip)), the `programs_js` extracted from the book, one file per snippet
 * Single-file Markdown export ([sicpjs.md](https://sicp.sourceacademy.org/sicpjs.md)), mirroring the PDF's content selection, for reading or feeding to tools that don't handle HTML/PDF/XML
 
@@ -18,7 +18,7 @@ This repository contains processing scripts and sources for *Structure and Inter
 * [PDF edition](https://sicp.sourceacademy.org/sicpy.pdf)
 * [Interactive SICPy](https://sourceacademy.org/sicpy), backed by structured per-section JSON textbook content, e.g. [json_py/1.1.1.json](https://sicp.sourceacademy.org/json_py/1.1.1.json)
 * [Comparison edition](https://sicp.sourceacademy.org/split_py), Scheme vs. Python
-* Search support: same mechanism as SICP JS above, via [json_py/rewritedSearchData.json](https://sicp.sourceacademy.org/json_py/rewritedSearchData.json)
+* Search support: same mechanism as SICP JS above, via [json_py/searchData.json](https://sicp.sourceacademy.org/json_py/searchData.json)
 * AI support: see [SICPy chatbot knowledge base](#sicpy-chatbot-knowledge-base) below
 * Source code archive ([sicpy.zip](https://sicp.sourceacademy.org/sicpy.zip)), the `programs_py` extracted from the book, one file per snippet
 * Single-file Markdown export ([sicpy.md](https://sicp.sourceacademy.org/sicpy.md)), mirroring the PDF's content selection
@@ -26,7 +26,7 @@ This repository contains processing scripts and sources for *Structure and Inter
 ### Scheme (the original)
 
 * [PDF edition](https://sicp.sourceacademy.org/sicp.pdf)
-* Structured per-section JSON textbook content, e.g. [json_scm/1.1.1.json](https://sicp.sourceacademy.org/json_scm/1.1.1.json), including generated search data ([json_scm/rewritedSearchData.json](https://sicp.sourceacademy.org/json_scm/rewritedSearchData.json)) — not yet wired up to a live interactive frontend, unlike JS/Python above
+* Structured per-section JSON textbook content, e.g. [json_scm/1.1.1.json](https://sicp.sourceacademy.org/json_scm/1.1.1.json), including generated search data ([json_scm/searchData.json](https://sicp.sourceacademy.org/json_scm/searchData.json)) — not yet wired up to a live interactive frontend, unlike JS/Python above
 * Source code archive ([sicp.zip](https://sicp.sourceacademy.org/sicp.zip)), the `programs_scm` extracted from the book, one file per snippet
 * Single-file Markdown export ([sicp.md](https://sicp.sourceacademy.org/sicp.md)), mirroring the PDF's content selection
